changed decimal to integer

+++ backend/db/migrations/20220630230354-create-room.js
@@ -50,7 +50,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DECIMAL(7,2),
+        type: Sequelize.INTEGER,
         allowNull: false,
       },
       createdAt: {
